@@ -3,12 +3,13 @@ package com.sf.struct.practice;
 public class DoubleCheckSingleton {
     private volatile static DoubleCheckSingleton instance;
 
-    private DoubleCheckSingleton(){}
+    private DoubleCheckSingleton() {
+    }
 
-    public static DoubleCheckSingleton getInstance(){
-        if(instance == null){
-            synchronized (DoubleCheckSingleton.class){
-                if(instance == null){
+    public static DoubleCheckSingleton getInstance() {
+        if (instance == null) {
+            synchronized (DoubleCheckSingleton.class) {
+                if (instance == null) {
                     instance = new DoubleCheckSingleton();
                 }
             }
